[WIP] bump versions of spark and play. Global handling of CORS
Sticking with old versions of spark is not good (other versions of akka and others stuffs implied).
Hence, it's also needed to change the version of play.

I've some problem decoding the pre-forged queries I had, but it might be a format issue.
Better checking with the compliance test, @xtordoir could you give me a hand on that?